Add MHU interface and adapter
The interface definition is the same as the mhu.h common interface
that is coming from the TF-M project. The adapter just forwards the
calls to the MHU wrapper.

The OpenAMP was not updated to use this new interface because it will be
deprecated and the Secure (TF-M) side implementation does not use the
mhu.h and wrapper, which would be required to make it work correctly.

The MHU driver and wrapper have some minor changes, mainly addressing
cppcheck issues and input parameter NULL checkings.

Signed-off-by: Bence Balogh <bence.balogh@arm.com>
Change-Id: Ia4d6e6e068aa67d86098dfd1da6fc99d73582373