Fix passing MbedTLS config in env-test deployment
The baremetal-tests suite of the env-test deployment defines the CMake
MBEDTLS_USER_CONFIG_FILE option but doesn't pass it to the MbedTLS
build as a C macro. Therefore whatever config is set for MbedTLS doesn't
actually have any effect on the build. Fix this by adding the missing
macro definition.

Signed-off-by: Balint Dobszay <balint.dobszay@arm.com>
Change-Id: I74bfaf37c8b219c4ac80dc1cd7dfba5d6182d6bc