Fix crypto discovery in standalone block-storage
The standalone block-storage service uses an RPMB block store as backend
which in turn uses the crypto service for key derivation operations.
Since commit abec4644 ("Enable owner labelling of keys in Mbed TLS") the
PSA Crypto implementation in MbedTLS cannot be called directly even in
the standalone service implementations, but the usual crypto service
discovery and a crypto client is needed instead. This was missing since
the commit mentioned above was merged after the standalone block-store
service was updated to use the RPMB block store, fix it now.

Signed-off-by: Balint Dobszay <balint.dobszay@arm.com>
Change-Id: Ifc41ef3a12f13d4cd547578b9dfa1300d377caa1