Create uefi test vectors with different timestamps
Update the test vector generator script to wait a bit between the
generation of variables to have different timestamps.
Change the order of the creation to match the order of the steps in
the relevant tests. Test append request for authenticated variables.
Also regenerate the test vectors:
E.g. db1 must be older than db2 to overwrite it.

Signed-off-by: Gabor Toth <gabor.toth2@arm.com>
Change-Id: I600cf0f4b7407ec947f8d7397c496ef63315c18b