Fix GetNextVariableName NameSize input
Based on the specification the NameSize shall be set to the available
buffer size at the first call instead of the NameSize of the
provided variable.
Change smm-gateway and the tests according this. Also remove
sanitize_get_next_var_name_param utility function, which is not
compilant with this solution.

Signed-off-by: Gabor Toth <gabor.toth2@arm.com>
Change-Id: I0c72e6d8cfc138991c169bd0510f0051fb691996