Fix endpoint ID in SPMC test
The SPMC test SP has a test case for checking the returned error for an
invalid direct request (where a2 is non-null, but also not a framework
message flag). However, this test case was using an invalid destination
endpoint ID, and was not actually testing the condition that it intended
to do. Fix this by using a valid destination for this message.

Signed-off-by: Balint Dobszay <balint.dobszay@arm.com>
Change-Id: I8340e4a7e4a935fc76747d41494da45810bc7212